First working version! Yay!
It works! (... at least on my computer). Assuming you have a recent enough version of python installed (coded in Python 3), place "main.py" and "Toggle Device (run as admin).bat" in the same folder and run the latter as admin to run the program. Needed due to elevated privilege requirements to turn devices on/off. Can stull run the script and see the summary, but will get ACCESS DENIED error when attempt to enable or disable anything.